+++ src/coordinax/_src/vectors/compat.py
@@ -125,11 +125,11 @@ def vec_diff_to_q(obj: QConvertible2D, /) -> Shaped[UncheckedQuantity, "*batch 2
 
     >>> vel = cx.vecs.CartesianVel2D.from_([1, 2], "km/s")
     >>> convert(vel, UncheckedQuantity)
-    UncheckedQuantity(Array([1., 2.], dtype=float32), unit='km / s')
+    UncheckedQuantity(Array([1, 2], dtype=int32), unit='km / s')
 
     >>> acc = cx.vecs.CartesianAcc2D.from_([1, 2], "km/s2")
     >>> convert(acc, UncheckedQuantity)
-    UncheckedQuantity(Array([1., 2.], dtype=float32), unit='km / s2')
+    UncheckedQuantity(Array([1, 2], dtype=int32), unit='km / s2')
 
     """
     return convert(_vec_diff_to_q(obj), UncheckedQuantity)
@@ -148,11 +148,11 @@ def vec_diff_to_q(obj: QConvertible2D, /) -> Shaped[Quantity, "*batch 2"]:
 
     >>> vel = cx.vecs.CartesianVel2D.from_([1, 2], "km/s")
     >>> convert(vel, u.Quantity)
-    Quantity['speed'](Array([1., 2.], dtype=float32), unit='km / s')
+    Quantity['speed'](Array([1, 2], dtype=int32), unit='km / s')
 
     >>> acc = cx.vecs.CartesianAcc2D.from_([1, 2], "km/s2")
     >>> convert(acc, u.Quantity)
-    Quantity['acceleration'](Array([1., 2.], dtype=float32), unit='km / s2')
+    Quantity['acceleration'](Array([1, 2], dtype=int32), unit='km / s2')
 
     """
     return convert(_vec_diff_to_q(obj), Quantity)
